Statusbar moves on mouseover fix
The floating results statusbar wasn't moving
out of the way when the mouse went over it,
making it difficult or impossible to see
some info on the last line of the results.
This is just a small fix to restore the
intended behavior. Also added a little padding
to make it evenly spaced on all sides, along with
a "frame" class so it now has a border in many
themes where it is currently missing one.