Use "match all" method for fulltext search (fixes #30)
This changes the default behavior of the fulltext
search to use the logical AND method for results
(this does not change the "Match results exactly" option).

Currently if searching for "bash" and "script"
you will get results for anything matching "bash" or "script"
(which can be done by searching for either word individually).

This change will reduce results as you add keywords. Results
will include files containing both "bash" AND "script" or
any part of both of those keywords ("bash" and "scripting",
"abashed" and "description").