Fix Thunar "Show in FM" bug, better FM support (Fixes #43)
This fixes a bug when right-clicking on results and
selecting "Show in File Manager" with Thunar.
Thunar wasn't highlighting files with spaces in
their names and would not open if a file if it
had certain characters in its name (like #).

Also now opens folder locations instead of folders
themselves. Previously "Open" and "Show in File Manager"
had the same results.

This also adds better support for more File Managers,
and support for detecting a custom File Manager if entered
in exo-preferred-applications. Nautilus and Thunar now
select files and folders. Nemo, and Elementary Files will
now select files when opened. Selecting folders with these
may be possible with dbus in the future. Fixed a problem
where Caja previously errored when trying to
"Show in File Manager" with files.