Added keyboard hotkeys for search entry, new open dialog
This adds keyboard accelerators for focusing
and highlighting the toolbar search entry
as well as bringing up the open folder dialog
for choosing a location. Should make things
better for accessibility and keyboard-centric
users in general.

A new GtkFileChooserDialog was added via
glade for this as hotkey would not apply
to the automatic File Chooser generated by
GtkFileChooserButton.

Added to the new File Chooser Dialog is a
warning if a folder isn't selected or
a manually entered folder is typed wrong.

Hotkeys:
Toolbar Search Entry = Ctrl+f
Open Folder Dialog = Ctrl+o or Ctrl+l