Adds symbolic link icons, better icon updating (fixes #46)
This adds a symbolic link emblem to symlink results.
It also rechecks icons when switching themes
(currently, when you switch icon themes the icons
don't switch like other apps).

The new icon method creates a new GdkPixmap that
is larger than the icon. It creates a 22x21 empty
image, composites the 16x16 icon, and then offsets
and composites the emblem on top of that.
Making the larger image allows for a nicer offset
for the symlink emblem.

Creating a Gio.EmblemedIcon (the 'proper' way)
created an icon with an emblem, but the emblem
covered the icon. This was is more convoluted
but gives better results, more similar to
Thunar and Xfdesktop.

Removed icon caching and overrides. Fallbacks
and extra checks should make up for them.