Sidebar background color bugfix
Currently the selectable rows of
the sidebar do not match the sidebar
background color in several Gtk themes
(the theme base color is used
for the sidebar rows).

This change makes the background of
the filetype and modified treeview
rows transparent so that only the
sidebar color shows.

Should improve compatibility and look
a little nicer on more themes.