Add symlink info in thumbnail view
Fixes issue where symlink icon would not show on
image thumbnails.

Adding a symlink to the image thumbnail would be
confusing visually (is it part of the image?),
and defeat the purpose of the thumbnailer to use
already existing thumbnail.

The solution here is to add a Symlink column.

Also added symlink target info to the details,
and tried to make the details a little more uniform.