Create shared filetype lists for searching and filtering
This creates a shared filetype list that both the Document
filetype filter, and fulltext search can use. This way
they stay in sync with any additions or changes.

The "text" list is used for both Documents filter and fulltext search
(this is a list of text-searchable file mimetypes), and the
"document" list is used for only the Document filter as these are
text documents that are technically binary (pdf, docx, etc) and
while they may only be searched using fulltext in some cases, they
should always show when filtering documents.

Fixes: #83