Ensure site-packages directory is prepended to sys.path
Otherwise, when the installation prefix is different from /usr and a
concurrent installation in /usr exists, it is used instead.
Unfortunately, prepending the prefix to site.PREFIXES does not prepend
the path to sys.path, it is appended. Not sure this is the best way to
do it, but it works.

Amends: 977e8c97f1029080bd22a6383ace74e0971a8da5